fix: keep argument pointers as pointers instead of laundering them through uintptr
packArg built the cell it hands to the FFI layer by storing an address in a
uintptr. A uintptr is an ordinary integer: the garbage collector does not scan
it and it keeps nothing alive. For the pointer, slice and array cases the only
thing standing between the callee and freed memory was the keepAlive slot, and
for the pointer/slice case that slot held the *uintptr cell rather than the
value it pointed at, so it kept the wrong object alive.

Storing an unsafe.Pointer in the cell instead fixes both halves at once: the
collector scans the cell and keeps the target alive for the duration of the
call, and the keepAlive slot now holds the value itself. The cell is still one
pointer wide, so nothing about the ABI changes. v.UnsafePointer() and
v.Addr().UnsafePointer() replace v.Pointer() and v.UnsafeAddr(), which are the
same numbers with the provenance kept.

getSymbolName did the same thing with the Dl_info struct it passes to dladdr.

The remaining conversion, cStringToGoString, is the opposite case and stays as
it is. Its address is a char* a native function returned; that memory is not
the collector's and never moves, so the conversion is correct and cannot be
written any other way. It is now marked //go:nocheckptr with the reason, which
also stops -race (which implies -d=checkptr) from aborting the process the
moment anything reads a returned string.

This does not make `go test -race ./...` pass on its own. What is left is not
in this repository:

  - goffi's ffi.callbackWrap (callback_arm64.go:189) has the same checkptr
    problem and is hit first on arm64
  - TestFloatAndBool fails under -race on main as well, five runs out of five
  - TestFullCircleFFI fails on main with and without -race

The float path and the callback trampoline were left alone deliberately: no
failure in this repository reaches them, and the callback address comes from
goffi rather than from Go's heap.

+++ sym_unix.go
@@ -54,8 +54,8 @@ func getSymbolName(addr uintptr) string {
 	var res int32
 
 	// goffi expects a pointer to the value that will be passed to C.
-	// Since C expects a pointer to the info struct, we must pass a pointer TO the uintptr containing the struct's address.
-	infoPtr := uintptr(unsafe.Pointer(&info))
+	// Since C expects a pointer to the info struct, we must pass a pointer TO the pointer containing the struct's address.
+	infoPtr := unsafe.Pointer(&info)
 
 	args := []unsafe.Pointer{
 		unsafe.Pointer(&addr),
